+++ static/js/settings.js
@@ -1,0 +1,15 @@
+define([], function(){
+  var Settings = {
+    languages: [
+      { lang: 'css', name: 'CSS' },
+      { lang: 'html', name: 'HTML' },
+      { lang: 'javascript', name: 'JavaScript' },
+      { lang: 'dom', name: 'DOM' },
+      { lang: 'jquery', name: 'jQuery' },
+      { lang: 'php', name: 'PHP' },
+      { lang: 'python', name: 'Python' },
+      { lang: 'python3', name: 'Python3' }
+    ]
+  }
+  return Settings;
+});