+++ static/js/views/topnav.js
@@ -37,7 +37,7 @@ define([
     },
 
     render: function() {
-      $(this.el).html(this.template({}));
+      $(this.el).html(this.template({ settings: this.options.settings }));
       return this;
     },
   });