+++ CHANGELOG.md
@@ -8,4 +8,6 @@ DocHub
   <br />[Remove bottom scrollbar](https://github.com/kevinburke/dochub/commit/2162b925b6ab2e3b9308966f73e1d7b859cfe044)
   <br />[Condense search window when there are fewer results ](https://github.com/kevinburke/dochub/commit/b4f29a213665e7a4d52eef3f74047071043f7a36)
   <br />[retina favicon](https://github.com/jamiethepiper/dochub/commit/ccb855db3fc935023e4af0526b4e190c779425ba)
-  <br />[typo](https://github.com/Fil/dochub/commit/09b3ca349b417222e38b059676afb46dc3d4abde) 
+  <br />[typo](https://github.com/Fil/dochub/commit/09b3ca349b417222e38b059676afb46dc3d4abde)
+  <br />[settings.js control](https://github.com/groovecoder/instacss/commit/38cdf05fc55aa6b7dc17c2ac3d79eb0f9ab0a175)
+* Added [Font-Awesome](https://github.com/FortAwesome/Font-Awesome)